The Hello World(part3) __init and __exit Macros
the __init macro causes the init function to be discarded and its memory freed once
the init functions finished for built-int drivers, but not loadable modules.
there is also an __initdata which works similarly to __init but for init vars rather
than funcs.
The __exit maro causes the omission of the function when the module is built into
the kernel, and like __exit, has no effect for loadable modules.

Signed-off-by: Francisco Pedraza <pacopedraza87@gmail.com>