Copying to and from the kernel
This is a little example on how to copy into and from
the Kernel.

Let's remember that:
1. For writing into user-space, the method copy_to_user() is
provided. It takes three parameters.
a) Destination memory address in the process's address space.
b) Source pointer in kernel-space.
c) Size in bytes of the data to copy.

2. For reading fro user-space, the method copy_from_user()
is analogous to copy_to_user(). The function reads from
the second parameter into the first paramter the number of
bytes specified in the third parameter.

Both of these functions returnthe number of bytes they failed
to copy on error. On sucess, they return zero. It is a standard
for the syscall to return -EFAULT in the case of such an error.

Signed-off-by: Francisco Pedraza <pacopedraza87@gmail.com>